+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [unreleased]
+
+* Re-organized tests
+* Fixed base64 decoding for `done_by`
+
 ## [3.0.0] - 2019-11-12
 
 * Dropped Python2 support